Create a moving-average playlist selector (#1125)
* Break out playlist selector logic
Pull the basic playlist selector into a separate file so there's a shared place to build alternates.

* Create a moving-average playlist selector
Define a variant of the standard playlist selector that calculates a moving average of bandwidth and uses that to select a playlist.
